refactor: split Google Photos upgrade routes
